devtools/server/tests/chrome/test_inspector-dead-nodes.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1121528
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1121528</title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gWalker = null;
let gDoc = null;

addAsyncTest(async function() {
  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  gDoc = doc;
  const inspector = await target.getFront("inspector");
  gWalker = inspector.walker;
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.children(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "body");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.children(nodeFront);
  await newRoot;

  ok(true, "The call to walker.children() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.nextSibling(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.nextSibling(nodeFront);
  await newRoot;

  ok(true, "The call to walker.nextSibling() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.previousSibling(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.previousSibling(nodeFront);
  await newRoot;

  ok(true, "The call to walker.previousSibling() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.addPseudoClassLock(nodeFront) before the load completes " +
    "shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.addPseudoClassLock(nodeFront, ":hover");
  await newRoot;

  ok(true, "The call to walker.addPseudoClassLock() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.removePseudoClassLock(nodeFront) before the load completes " +
    "shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.removePseudoClassLock(nodeFront, ":hover");
  await newRoot;

  ok(true, "The call to walker.removePseudoClassLock() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.clearPseudoClassLocks(nodeFront) before the load completes " +
    "shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.clearPseudoClassLocks(nodeFront);
  await newRoot;

  ok(true, "The call to walker.clearPseudoClassLocks() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.innerHTML(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.innerHTML(nodeFront);
  await newRoot;

  ok(true, "The call to walker.innerHTML() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.setInnerHTML(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.setInnerHTML(nodeFront, "<span>innerHTML changed</span>");
  await newRoot;

  ok(true, "The call to walker.setInnerHTML() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.outerHTML(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.outerHTML(nodeFront);
  await newRoot;

  ok(true, "The call to walker.outerHTML() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.setOuterHTML(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.setOuterHTML(nodeFront, "<h1><span>innerHTML changed</span></h1>");
  await newRoot;

  ok(true, "The call to walker.setOuterHTML() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.insertAdjacentHTML(nodeFront) before the load completes shouldn't " +
    "fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.insertAdjacentHTML(nodeFront, "afterEnd",
    "<span>new adjacent HTML</span>");
  await newRoot;

  ok(true, "The call to walker.insertAdjacentHTML() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.removeNode(nodeFront) before the load completes should throw");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  let hasThrown = false;
  try {
    await gWalker.removeNode(nodeFront);
  } catch (e) {
    hasThrown = true;
  }
  await newRoot;

  ok(hasThrown, "The call to walker.removeNode() threw");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.removeNodes([nodeFront]) before the load completes should throw");

  const nodeFront1 = await gWalker.querySelector(gWalker.rootNode, "h1");
  const nodeFront2 = await gWalker.querySelector(gWalker.rootNode, "#longstring");
  const nodeFront3 = await gWalker.querySelector(gWalker.rootNode, "#shortstring");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  let hasThrown = false;
  try {
    await gWalker.removeNodes([nodeFront1, nodeFront2, nodeFront3]);
  } catch (e) {
    hasThrown = true;
  }
  await newRoot;

  ok(hasThrown, "The call to walker.removeNodes() threw");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.insertBefore(nodeFront, parent, null) before the load completes " +
    "shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newParentFront = await gWalker.querySelector(gWalker.rootNode, "#longlist");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.insertBefore(nodeFront, newParentFront);
  await newRoot;

  ok(true, "The call to walker.insertBefore() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.insertBefore(nodeFront, parent, sibling) before the load completes " +
    "shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newParentFront = await gWalker.querySelector(gWalker.rootNode, "#longlist");
  const siblingFront = await gWalker.querySelector(gWalker.rootNode, "#b");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.insertBefore(nodeFront, newParentFront, siblingFront);
  await newRoot;

  ok(true, "The call to walker.insertBefore() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.editTagName(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.editTagName(nodeFront, "h2");
  await newRoot;

  ok(true, "The call to walker.editTagName() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.hideNode(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.hideNode(nodeFront);
  await newRoot;

  ok(true, "The call to walker.hideNode() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.unhideNode(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.unhideNode(nodeFront);
  await newRoot;

  ok(true, "The call to walker.unhideNode() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.releaseNode(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "h1");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.releaseNode(nodeFront);
  await newRoot;

  ok(true, "The call to walker.releaseNode() didn't fail");
  runNextTest();
});

addAsyncTest(async function() {
  info("Getting a nodeFront, reloading the page, and calling " +
    "walker.querySelector(nodeFront) before the load completes shouldn't fail");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "body");
  const newRoot = waitForMutation(gWalker, isNewRoot);
  gDoc.defaultView.location.reload();
  await gWalker.querySelector(nodeFront, "h1");
  await newRoot;

  ok(true, "The call to walker.querySelector() didn't fail");
  runNextTest();
});

addTest(function cleanup() {
  gWalker = null;
  gDoc = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=932937">Mozilla Bug 1121528</a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
